Factor out add_system_constraint to dynamic_system.py
