+++ neural_network_lyapunov/lyapunov.py
@@ -13,6 +13,7 @@ import neural_network_lyapunov.feedback_system as feedback_system
 import neural_network_lyapunov.gurobi_torch_mip as gurobi_torch_mip
 import neural_network_lyapunov.utils as utils
 import neural_network_lyapunov.mip_utils as mip_utils
+import neural_network_lyapunov.dynamic_system as dynamic_system
 
 
 class ConvergenceEps(Enum):
@@ -59,55 +60,6 @@ class LyapunovHybridLinearSystem:
                 lyapunov_relu, self.system.dtype)
         self.network_bound_propagate_method = \
             mip_utils.PropagateBoundsMethod.IA
-
-    def add_system_constraint(self,
-                              milp,
-                              x,
-                              x_next,
-                              *,
-                              binary_var_type=gurobipy.GRB.BINARY):
-        """
-        This function is intended for internal usage only (but I expose it
-        as a public function for unit test).
-        Add the constraint and variables to write the hybrid linear system
-        dynamics as mixed-integer linear constraints.
-        @param binary_var_type Refer to GurobiTorchMIP.addVars for more
-        details.
-        """
-        if isinstance(
-            self.system, hybrid_linear_system.AutonomousHybridLinearSystem)\
-                or isinstance(self.system, relu_system.AutonomousReLUSystem)\
-                or isinstance(
-                    self.system,
-                    relu_system.AutonomousReLUSystemGivenEquilibrium)\
-                or isinstance(
-                    self.system,
-                    relu_system.AutonomousResidualReLUSystemGivenEquilibrium):
-            assert (isinstance(milp, gurobi_torch_mip.GurobiTorchMIP))
-            return self.system.add_dynamics_constraint(milp, x, x_next, "s",
-                                                       "gamma",
-                                                       binary_var_type)
-
-        elif isinstance(self.system, feedback_system.FeedbackSystem):
-            u, forward_dynamics_return, controller_mip_cnstr_return = \
-                self.system.add_dynamics_mip_constraint(
-                    milp, x, x_next, "u", "forward_s", "forward_binary",
-                    "controller_s", "controller_binary", binary_var_type)
-            slack = u + forward_dynamics_return.slack +\
-                controller_mip_cnstr_return.slack
-            binary = forward_dynamics_return.binary + \
-                controller_mip_cnstr_return.binary
-            ret = hybrid_linear_system.DynamicsConstraintReturn(slack, binary)
-            ret.x_next_lb_IA = forward_dynamics_return.x_next_lb_IA
-            ret.x_next_ub_IA = forward_dynamics_return.x_next_ub_IA
-            ret.x_next_bound_prog = forward_dynamics_return.x_next_bound_prog
-            ret.x_next_bound_var = forward_dynamics_return.x_next_bound_var
-            ret.x_var = forward_dynamics_return.x_var
-            ret.forward_dynamics_return = forward_dynamics_return
-            ret.controller_mip_cnstr_return = controller_mip_cnstr_return
-            return ret
-        else:
-            raise (NotImplementedError)
 
     def add_lyap_relu_output_constraint(self,
                                         milp,
@@ -702,8 +654,8 @@ class LyapunovDiscreteTimeHybridSystem(LyapunovHybridLinearSystem):
                          name="x")
 
         # x is the variable x[n]
-        system_constraint_return = self.add_system_constraint(
-            milp, x, x_next, binary_var_type=binary_var_type)
+        system_constraint_return = dynamic_system._add_system_constraint(
+            self.system, milp, x, x_next, binary_var_type=binary_var_type)
         s = system_constraint_return.slack
         gamma = system_constraint_return.binary
         # warmstart the binary variables
@@ -1200,7 +1152,8 @@ class LyapunovDiscreteTimeHybridSystem(LyapunovHybridLinearSystem):
                               name="x[n+1]")
 
         # Add the dynamics constraint between x[n+1] and x[n]
-        self.add_system_constraint(milp, x_curr, x_next)
+        dynamic_system._add_system_constraint(
+            self.system, milp, x_curr, x_next)
 
         if x_curr_in_box:
             out_box_x = x_next
